# docs/swagger.yaml
basePath: /
definitions:
...
  models.DeleteMediaRequest:
    properties:
      media_id:
        type: string
    required:
    - media_id
    type: object
  models.DeleteMediaResponse:
    properties:
      media_id:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  models.GetLinkListRequest:
    properties:
      content_id:
        type: string
    required:
    - content_id
    type: object
  models.GetLinkMediaRequest:
    properties:
      media_id:
        type: string
    required:
    - media_id
    type: object
  models.GetLinkMediaURIRequest:
    properties:
      expiry_minutes:
        type: integer
      media_id:
        type: string
    required:
    - media_id
    type: object
  models.GetLinkMediaURIResponse:
    properties:
      expires_at:
        type: string
      media_id:
        type: string
      s3_url:
        type: string
    type: object
...
  models.MediaListItem:
    properties:
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      media_id:
        type: string
      upload_date:
        type: string
    type: object
  models.MediaListResponse:
    properties:
      content_id:
        type: string
      link:
        type: string
      media_files:
        items:
          $ref: '#/definitions/models.MediaListItem'
        type: array
    type: object
...
  models.UpdateMediaRequest:
    properties:
      file_name:
        type: string
      media_id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      original_file_name:
        type: string
    required:
    - media_id
    type: object
  models.UpdateMediaResponse:
    properties:
      media_id:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Go-based microservice for planning and scheduling YouTube live streams.
    Helps content creators organize their streaming schedule and manage content.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: St. Planer - YouTube Stream Planner API
  version: "1.0"
paths:
...
  /api/v1/media/get:
    delete:
      consumes:
      - application/json
      description: Delete media file from database and S3 storage
      parameters:
      - description: Media delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteMediaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete media file
      tags:
      - media
    post:
      consumes:
      - application/json
      description: Download specific media file from a post as binary stream. Supports
        range requests for video files to enable streaming and seeking.
      parameters:
      - description: Media download request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GetLinkMediaRequest'
      - description: Range header for partial content (e.g., bytes=0-1023)
        in: header
        name: Range
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Full file download
          schema:
            type: file
        "206":
          description: Partial content (range request)
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "416":
          description: Range Not Satisfiable
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Download specific media file
      tags:
      - media
    put:
      consumes:
      - application/json
      description: Update media file metadata including filename and custom metadata
      parameters:
      - description: Media update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateMediaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update media metadata
      tags:
      - media
  /api/v1/media/getDirect:
    post:
      consumes:
      - application/json
      description: Get direct S3 link for specific media with configurable expiration
      parameters:
      - description: Media URI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GetLinkMediaURIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetLinkMediaURIResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get S3 pre-signed URL for media
      tags:
      - media
...
  /api/v1/media/links:
    post:
      consumes:
      - application/json
      description: Get list of all media files from a specific Telegram post or YouTube
        video
      parameters:
      - description: Content ID for post
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GetLinkListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MediaListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get media files from a specific post
      tags:
      - media
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"